Add slice folder and moved CreateJobSlice in there and change paths
